back/pyproject.toml
[tool.poetry]
name = "flex2"
version = "0.1.0"
description = ""
authors = ["connbrack"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0"
geopy = "^2.4"
numpy = "^1.26"
playwright = "^1.43"
python-dotenv = "^1.0"
requests = "^2.31"
flask-cors = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
